+++ app/controllers/Cargos.java
@@ -1,6 +1,8 @@
 package controllers;
 
+import models.Area;
 import models.Cargo;
+import models.Vaga;
 import play.data.Form;
 import play.mvc.Controller;
 import play.mvc.Result;
@@ -9,13 +11,20 @@ import com.avaje.ebean.Page;
 
 public class Cargos extends Controller{
 	private static final Form<Cargo> cargoForm = Form.form(Cargo.class);
+	public static Result GO_HOME = redirect(
+	        routes.Cargos.list(0, "id", "asc", "", "nome")
+	    );
+	public static Result list(int page, String sortBy, String order, String filter, String atributo) {
+        return ok(
+        		views.html.cargos.list.render(
+            	Cargo.page(page, 10, sortBy, order, filter, atributo),
+                sortBy, order, filter, atributo
+            )
+        );
+    }
 	public static Result index() {
-	    return redirect(routes.Cargos.lista(0));
-	  }
-	public static Result lista(Integer page){
-		Page<Cargo> cargos = Cargo.buscarTodos(page);
-	    return ok(views.html.cargos.list.render(cargos));
-	}
+        return GO_HOME;
+    }
 	public static Result novo(){
 		 return ok(views.html.cargos.detalhes.render(cargoForm));
 	}
@@ -37,14 +46,18 @@ public class Cargos extends Controller{
 	    }
 		flash("success",
 		        String.format("Cargo atualizado %s", cargo));
-	    return redirect(routes.Cargos.lista(0));
+	    return redirect(routes.Cargos.list(0, "id", "asc", "", "nome"));
 	}
 	public static Result delete(Long id) {
 	  final Cargo cargo = Cargo.buscarPorId(id);
 	  if(cargo == null) {
 	    return notFound(String.format("Cargo com id %s não existe.", id));
 	  }
+	  final Vaga vaga = Vaga.buscarPorCargo(id);
+	  if(vaga != null) {
+	    return badRequest(String.format("Cargo %s não pode ser excluído.", id));
+	  }
 	  cargo.delete();
-	  return redirect(routes.Cargos.lista(0));
+	  return redirect(routes.Cargos.list(0, "id", "asc", "", "nome"));
 	}
 }
